+++ backend/canisters/user_index/api/src/queries/mod.rs
@@ -9,6 +9,7 @@ pub mod current_user;
 pub mod diamond_membership_fees;
 pub mod explore_bots;
 pub mod external_achievements;
+pub mod migrated_user_ids;
 pub mod moderation_config;
 pub mod platform_moderators;
 pub mod platform_moderators_group;